Fix max valid time (#15)
