+++ src/main/java/com/codility/max_valid_time/Main.java
@@ -12,7 +12,7 @@ public class Main
     
     int[][] testcase = getTestCase(); 
 
-    Solution sol = new Solution02v2();
+    Solution sol = new Solution02();
     
 
     for (int[] nums : testcase)
