--- src/main/java/com/example/Maximum_valid_time/Solution.java
@@ -1,6 +1,0 @@
-package com.example.Maximum_valid_time;
-
-public interface Solution
-{
-  public String solution(int A, int B, int C, int D); 
-}
